+++ back-end/package.json
@@ -14,6 +14,7 @@
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "path": "^0.12.7",
+    "save": "^2.9.0",
     "spotify-web-api-node": "^5.0.0"
   },
   "devDependencies": {
